Moved events to a new subproject: paper
